+++ src/domain/mod.rs
@@ -1,6 +1,7 @@
 pub mod osm_node;
 pub mod osm_pbf_file;
 pub mod osm_relation;
+pub mod osm_tag;
 pub mod osm_way;
 
 pub(crate) trait table {